requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.3
omegaconf==2.3.0
openai==1.67.0
pandas==2.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
rich==13.9.4
rootutils==1.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1 
